enable sbom
