+++ app/controlplane/cmd/wire.go
@@ -38,7 +38,6 @@ func wireApp(*conf.Bootstrap, credentials.ReaderWriter, log.Logger) (*app, func(
 	panic(
 		wire.Build(
 			wire.Bind(new(credentials.Reader), new(credentials.ReaderWriter)),
-			wire.Bind(new(credentials.Writer), new(credentials.ReaderWriter)),
 			server.ProviderSet,
 			data.ProviderSet,
 			biz.ProviderSet,
